requirements.txt
fastapi
uvicorn
streamlit
streamlit-mic-recorder
langchain
langchain-community
langchain-ollama
langchain-huggingface
python-multipart
chromadb
sentence-transformers
google-search-results
transformers
torch
pypdf
google-generativeai